+++ server/src/internalClusterTest/java/org/opensearch/remotestore/RemoteStoreRestoreIT.java
@@ -15,13 +15,11 @@ import org.opensearch.action.support.PlainActionFuture;
 import org.opensearch.cluster.health.ClusterHealthStatus;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.unit.TimeValue;
-import org.opensearch.core.common.unit.ByteSizeUnit;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.test.InternalTestCluster;
 import org.opensearch.test.OpenSearchIntegTestCase;
 import org.opensearch.test.transport.MockTransportService;
-import org.junit.Before;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -35,7 +33,7 @@ import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertAcked;
 import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertHitCount;
 import static org.hamcrest.Matchers.greaterThan;
 
-@OpenSearchIntegTestCase.ClusterScope(scope = OpenSearchIntegTestCase.Scope.SUITE, numDataNodes = 0)
+@OpenSearchIntegTestCase.ClusterScope(scope = OpenSearchIntegTestCase.Scope.TEST, numDataNodes = 0)
 public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
     private static final String INDEX_NAME = "remote-store-test-idx-1";
     private static final String INDEX_NAMES = "test-remote-store-1,test-remote-store-2,remote-store-test-index-1,remote-store-test-index-2";
@@ -51,11 +49,6 @@ public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
     @Override
     protected Collection<Class<? extends Plugin>> nodePlugins() {
         return Arrays.asList(MockTransportService.TestPlugin.class);
-    }
-
-    @Before
-    public void setup() {
-        setupRepo();
     }
 
     private void restore(String... indices) {
@@ -94,8 +87,19 @@ public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
     }
 
     private void prepareCluster(int numClusterManagerNodes, int numDataOnlyNodes, String indices, int replicaCount, int shardCount) {
-        internalCluster().startClusterManagerOnlyNodes(numClusterManagerNodes);
-        internalCluster().startDataOnlyNodes(numDataOnlyNodes);
+        prepareCluster(numClusterManagerNodes, numDataOnlyNodes, indices, replicaCount, shardCount, Settings.EMPTY);
+    }
+
+    private void prepareCluster(
+        int numClusterManagerNodes,
+        int numDataOnlyNodes,
+        String indices,
+        int replicaCount,
+        int shardCount,
+        Settings clusterSettings
+    ) {
+        internalCluster().startClusterManagerOnlyNodes(numClusterManagerNodes, clusterSettings);
+        internalCluster().startDataOnlyNodes(numDataOnlyNodes, clusterSettings);
         for (String index : indices.split(",")) {
             createIndex(index, remoteStoreIndexSettings(replicaCount, shardCount));
             ensureYellowAndNoInitializingShards(index);
@@ -187,7 +191,7 @@ public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
      * @throws IOException IO Exception.
      */
     private void testRestoreFlow(int numberOfIterations, boolean invokeFlush, int shardCount) throws Exception {
-        prepareCluster(0, 3, INDEX_NAME, 0, shardCount);
+        prepareCluster(1, 3, INDEX_NAME, 0, shardCount);
         Map<String, Long> indexStats = indexData(numberOfIterations, invokeFlush, INDEX_NAME);
         assertEquals(shardCount, getNumShards(INDEX_NAME).totalNumShards);
 
@@ -269,7 +273,7 @@ public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
 
     public void testRestoreFlowAllShardsNoRedIndex() throws InterruptedException {
         int shardCount = randomIntBetween(1, 5);
-        prepareCluster(0, 3, INDEX_NAME, 0, shardCount);
+        prepareCluster(1, 3, INDEX_NAME, 0, shardCount);
         indexData(randomIntBetween(2, 5), true, INDEX_NAME);
         assertEquals(shardCount, getNumShards(INDEX_NAME).totalNumShards);
 
@@ -285,7 +289,7 @@ public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
 
     public void testRestoreFlowNoRedIndex() throws Exception {
         int shardCount = randomIntBetween(1, 5);
-        prepareCluster(0, 3, INDEX_NAME, 0, shardCount);
+        prepareCluster(1, 3, INDEX_NAME, 0, shardCount);
         Map<String, Long> indexStats = indexData(randomIntBetween(2, 5), true, INDEX_NAME);
         assertEquals(shardCount, getNumShards(INDEX_NAME).totalNumShards);
 
@@ -459,22 +463,16 @@ public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
     }
 
     public void testRateLimitedRemoteDownloads() throws Exception {
-        assertAcked(
-            client().admin()
-                .cluster()
-                .preparePutRepository(REPOSITORY_NAME)
-                .setType("fs")
-                .setSettings(
-                    Settings.builder()
-                        .put("location", randomRepoPath())
-                        .put("compress", randomBoolean())
-                        .put("max_remote_download_bytes_per_sec", "2kb")
-                        .put("chunk_size", 200, ByteSizeUnit.BYTES)
-
-                )
-        );
+        clusterSettingsSuppliedByTest = true;
         int shardCount = randomIntBetween(1, 3);
-        prepareCluster(0, 3, INDEX_NAME, 0, shardCount);
+        prepareCluster(
+            1,
+            3,
+            INDEX_NAME,
+            0,
+            shardCount,
+            buildRemoteStoreNodeAttributes(REPOSITORY_NAME, randomRepoPath(), REPOSITORY_2_NAME, randomRepoPath(), true)
+        );
         Map<String, Long> indexStats = indexData(5, false, INDEX_NAME);
         assertEquals(shardCount, getNumShards(INDEX_NAME).totalNumShards);
         internalCluster().stopRandomNode(InternalTestCluster.nameFilter(primaryNodeName(INDEX_NAME)));